Add menu option to update bookmark It updates bookmark with the current view setting. Doesn't save automatically yet
