+++ src/features/bookmarks/bookmark.tsx
@@ -1,4 +1,4 @@
-import { Delete, Edit, MoreVert, Share } from "@mui/icons-material";
+import { Cached, Delete, Edit, MoreVert, Share } from "@mui/icons-material";
 import {
     Box,
     IconButton,
@@ -18,13 +18,16 @@ import { css } from "@mui/styled-engine";
 import { MouseEvent, useState } from "react";
 import { useHistory } from "react-router-dom";
 
-import { useAppSelector } from "app/redux-store-interactions";
+import { useAppDispatch, useAppSelector } from "app/redux-store-interactions";
 import { Tooltip } from "components";
+import { useExplorerGlobals } from "contexts/explorerGlobals";
 import { selectUser } from "slices/authSlice";
 import { selectHasAdminCapabilities } from "slices/explorer";
 
-import { BookmarkAccess, ExtendedBookmark } from "./bookmarksSlice";
+import { BookmarkAccess, bookmarksActions, ExtendedBookmark, selectBookmarks } from "./bookmarksSlice";
+import { useCreateBookmark } from "./useCreateBookmark";
 import { useSelectBookmark } from "./useSelectBookmark";
+import { createBookmarkImg } from "./utils";
 
 const Description = styled(Typography)(
     () => css`
@@ -61,10 +64,16 @@ const Img = styled("img")(
 );
 
 export function Bookmark({ bookmark }: { bookmark: ExtendedBookmark }) {
+    const {
+        state: { canvas },
+    } = useExplorerGlobals(true);
     const theme = useTheme();
     const history = useHistory();
     const [menuAnchor, setMenuAnchor] = useState<HTMLElement | null>(null);
     const selectBookmark = useSelectBookmark();
+    const createBookmark = useCreateBookmark();
+    const dispatch = useAppDispatch();
+    const bookmarks = useAppSelector(selectBookmarks);
 
     const isAdmin = useAppSelector(selectHasAdminCapabilities);
     const user = useAppSelector(selectUser);
@@ -76,6 +85,15 @@ export function Bookmark({ bookmark }: { bookmark: ExtendedBookmark }) {
 
     const closeMenu = () => {
         setMenuAnchor(null);
+    };
+
+    const handleUpdate = async () => {
+        const { img, explorerState } = createBookmark(await createBookmarkImg(canvas));
+
+        const newBookmarks = bookmarks.map((bm) => (bm === bookmark ? { ...bm, img, explorerState } : bm));
+
+        dispatch(bookmarksActions.setBookmarks(newBookmarks));
+        closeMenu();
     };
 
     return (
@@ -171,6 +189,12 @@ export function Bookmark({ bookmark }: { bookmark: ExtendedBookmark }) {
                             <Edit fontSize="small" />
                         </ListItemIcon>
                         <ListItemText>Edit</ListItemText>
+                    </MenuItem>,
+                    <MenuItem key="update" onClick={handleUpdate}>
+                        <ListItemIcon>
+                            <Cached fontSize="small" />
+                        </ListItemIcon>
+                        <ListItemText>Update</ListItemText>
                     </MenuItem>,
                     <MenuItem key="delete" onClick={() => history.push(`delete/${bookmark.id}`)}>
                         <ListItemIcon>
